+++ lib/gg_one.dart
@@ -23,17 +23,15 @@ export 'package:gg_lang/gg_lang.dart'
 // Can
 export 'package:gg_one_commit/gg_one_commit.dart';
 export 'src/commands/can.dart';
-export 'src/commands/can/can_publish.dart';
 // Checks
 export 'package:gg_one_checks/gg_one_checks.dart';
 // Did
 export 'src/commands/did.dart';
-export 'src/commands/did/did_publish.dart';
 export 'src/commands/did/did_upgrade.dart';
 // Do
+export 'package:gg_one_do_publish/gg_one_do_publish.dart';
 export 'src/commands/do.dart';
 export 'src/commands/do/create.dart';
-export 'src/commands/do/do_publish.dart';
 export 'src/commands/do/do_upgrade.dart';
 // Info
 export 'src/commands/info.dart';
@@ -42,6 +40,3 @@ export 'src/gg.dart';
 export 'package:gg_one_merge/gg_one_merge.dart';
 export 'package:gg_one_publish_config/gg_one_publish_config.dart';
 export 'package:gg_one_state/gg_one_state.dart';
-export 'src/tools/add_git_only_version_tag.dart';
-export 'src/tools/add_typescript_version_tag.dart';
-export 'src/tools/workspace_folder_guard.dart';